+++ .eslintrc.js
@@ -2,10 +2,8 @@ module.exports = {
   env: {
     browser: true
   },
-  extends: ['airbnb-base', 'prettier'],
-  plugins: ['prettier'],
+  extends: ['airbnb-base', 'plugin:prettier/recommended'],
   rules: {
-    'prettier/prettier': ['error'],
     'no-console': [
       'error',
       {